backend/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
APScheduler
attrs
backtesting
blinker
bokeh
btest
certifi
charset-normalizer
click
contourpy
dateparser
dill
fastapi
Flask
flask_socketio
flask_cors
cachetools
frozenlist
greenlet
h11
httpcore
httpx
idna
itsdangerous
Jinja2
lxml
MarkupSafe
multidict
multiprocess
multitasking
narwhals
numpy
packaging
pandas
passlib
pillow
pip
propcache
psutil
pycryptodome
pydantic
pydantic_core
python-binance
python_bitvavo_api
python-dateutil
python-dotenv
pytz
PyYAML
regex
requests
setuptools
six
sniffio
SQLAlchemy
starlette
tornado
typing_extensions
tzdata
tzlocal
urllib3
uvicorn
websocket-client
websockets
Werkzeug
xyzservices
yarl